# apps/mas-ops-api/openapi.json
{
  "components": {
    "schemas": {
      "ActivityEventResponse": {
        "description": "Incident activity timeline entry.",
        "properties": {
          "activity_id": {
            "title": "Activity Id",
            "type": "integer"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "fabric_id": {
            "title": "Fabric Id",
            "type": "string"
          },
          "incident_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "source_event_id": {
            "title": "Source Event Id",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "title": "Subject Type",
            "type": "string"
          }
        },
        "required": [
          "activity_id",
          "source_event_id",
          "client_id",
          "fabric_id",
          "incident_id",
          "asset_id",
          "event_type",
          "subject_type",
          "subject_id",
          "payload",
          "occurred_at"
        ],
        "title": "ActivityEventResponse",
        "type": "object"
      },
...
      "ConfigApplyResult": {
        "additionalProperties": false,
        "description": "Desired-state apply lifecycle result.",
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "config_apply_run_id": {
            "title": "Config Apply Run Id",
            "type": "string"
          },
          "desired_state_version": {
            "exclusiveMinimum": 0.0,
            "title": "Desired State Version",
            "type": "integer"
          },
          "error_summary": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ConfigApplyState"
          }
        },
        "required": [
          "config_apply_run_id",
          "client_id",
          "desired_state_version",
          "status"
        ],
        "title": "ConfigApplyResult",
        "type": "object"
      },
      "ConfigApplyState": {
        "description": "Canonical config validation and apply states.",
        "enum": [
          "pending",
          "validating",
          "applying",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "ConfigApplyState",
        "type": "string"
      },
      "ConfigDesiredState-Input": {
        "additionalProperties": false,
        "description": "Declarative configuration record.",
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "desired_state_version": {
            "exclusiveMinimum": 0.0,
            "title": "Desired State Version",
            "type": "integer"
          },
          "fabric_id": {
            "title": "Fabric Id",
            "type": "string"
          },
          "inventory_sources": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JSONValue-Input"
              },
              "type": "object"
            },
            "title": "Inventory Sources",
            "type": "array"
          },
          "notification_routes": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JSONValue-Input"
              },
              "type": "object"
            },
            "title": "Notification Routes",
            "type": "array"
          },
          "policy": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Input"
            },
            "title": "Policy",
            "type": "object"
          },
          "tenant_metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Input"
            },
            "title": "Tenant Metadata",
            "type": "object"
          }
        },
        "required": [
          "client_id",
          "fabric_id",
          "desired_state_version"
        ],
        "title": "ConfigDesiredState",
        "type": "object"
      },
      "ConfigDesiredState-Output": {
        "additionalProperties": false,
        "description": "Declarative configuration record.",
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "desired_state_version": {
            "exclusiveMinimum": 0.0,
            "title": "Desired State Version",
            "type": "integer"
          },
          "fabric_id": {
            "title": "Fabric Id",
            "type": "string"
          },
          "inventory_sources": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JSONValue-Output"
              },
              "type": "object"
            },
            "title": "Inventory Sources",
            "type": "array"
          },
          "notification_routes": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JSONValue-Output"
              },
              "type": "object"
            },
            "title": "Notification Routes",
            "type": "array"
          },
          "policy": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "title": "Policy",
            "type": "object"
          },
          "tenant_metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "title": "Tenant Metadata",
            "type": "object"
          }
        },
        "required": [
          "client_id",
          "fabric_id",
          "desired_state_version"
        ],
        "title": "ConfigDesiredState",
        "type": "object"
      },
      "ConfigValidationResult": {
        "additionalProperties": false,
        "description": "Validation output for a desired-state configuration.",
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "config_apply_run_id": {
            "title": "Config Apply Run Id",
            "type": "string"
          },
          "desired_state_version": {
            "exclusiveMinimum": 0.0,
            "title": "Desired State Version",
            "type": "integer"
          },
          "errors": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "status": {
            "minLength": 1,
            "title": "Status",
            "type": "string"
          },
          "validated_at": {
            "title": "Validated At",
            "type": "string"
          },
          "warnings": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "config_apply_run_id",
          "client_id",
          "desired_state_version",
          "status",
          "validated_at"
        ],
        "title": "ConfigValidationResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncidentResponse": {
        "description": "Incident detail response.",
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "fabric_id": {
            "title": "Fabric Id",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "client_id",
          "fabric_id",
          "state",
          "severity",
          "summary",
          "opened_at",
          "updated_at"
        ],
        "title": "IncidentResponse",
        "type": "object"
      },
      "JSONValue-Input": {
        "anyOf": [
          {
            "type": "boolean"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "string"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JSONValue-Input"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Input"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
      "JSONValue-Output": {
        "anyOf": [
          {
            "type": "boolean"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "string"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue-Output"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MAS Ops API",
    "version": "0.4.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/clients/{client_id}/config/apply": {
      "post": {
        "description": "Start an apply run for the latest desired state.",
        "operationId": "apply_desired_state_clients__client_id__config_apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigApplyResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Desired State",
        "tags": [
          "config"
        ]
      }
    },
    "/clients/{client_id}/config/desired-state": {
      "get": {
        "description": "Fetch the current desired-state config for one client.",
        "operationId": "get_desired_state_clients__client_id__config_desired_state_get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigDesiredState-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Desired State",
        "tags": [
          "config"
        ]
      },
      "put": {
        "description": "Replace the desired-state document for one client.",
        "operationId": "replace_desired_state_clients__client_id__config_desired_state_put",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigDesiredState-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigDesiredState-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Desired State",
        "tags": [
          "config"
        ]
      }
    },
    "/clients/{client_id}/config/validate": {
      "post": {
        "description": "Start a validation run for the latest desired state.",
        "operationId": "validate_desired_state_clients__client_id__config_validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Desired State",
        "tags": [
          "config"
        ]
      }
    },
...
    "/incidents/{incident_id}": {
      "get": {
        "description": "Fetch one incident detail.",
        "operationId": "get_incident_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "incidents"
        ]
      }
    },
    "/incidents/{incident_id}/activity": {
      "get": {
        "description": "Fetch activity timeline entries for one incident.",
        "operationId": "list_incident_activity_incidents__incident_id__activity_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActivityEventResponse"
                  },
                  "title": "Response List Incident Activity Incidents  Incident Id  Activity Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Incident Activity",
        "tags": [
          "incidents"
        ]
      }
    },
...
  }
}